requirements.txt
# Core exchange integration
ccxt>=4.0.0

# Async operations
aiohttp>=3.9.0

# Slack integration
slack-sdk>=3.23.0
slack-bolt>=1.18.0  # For Slack bot conversational interface

# Configuration management
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0  # Excel file support

# HTTP requests
requests>=2.31.0
httpx>=0.25.0  # Modern async HTTP client

# Sui Blockchain
pysui>=0.65.0  # Sui Python SDK

# Claude AI Analysis
anthropic>=0.39.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL driver
sqlalchemy>=2.0.0  # ORM for PostgreSQL

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
